refact: abstracting so we can test
